# allocate less in NewBucketWithRateAndClock