As suggested by @helinbo2015 in issue #19.

    benchmark                old ns/op     new ns/op     delta
    BenchmarkNewBucket-4     2965          892           -69.92%